Update image detail and styles in UI
The code updates have modified the UI to adjust image detail to "low" in the openai.ts file. In the page.tsx file, various styling modifications are done including the size of the display text, button design for image upload and other details. This commit aims to make the user interface more responsive and efficient.